+++ ExampleProject/platforms/ios/ExampleProject/Plugins/clevertap-cordova/CleverTapSDK.framework/PrivateHeaders/CTDeviceInfo.h
@@ -1,6 +1,7 @@
 #import <Foundation/Foundation.h>
 
 @class CleverTapInstanceConfig;
+@class CTValidationResult;
 
 @interface CTDeviceInfo : NSObject
 
@@ -21,11 +22,17 @@
 @property (strong, readonly) NSString *deviceWidth;
 @property (strong, readonly) NSString *deviceHeight;
 @property (atomic, readonly) NSString *deviceId;
+@property (atomic, readonly) NSString *deviceName;
+@property (atomic, readonly) NSString *fallbackDeviceId;
+@property (atomic, readwrite) NSString *library;
 @property (assign, readonly) BOOL wifi;
 @property (assign, readonly) BOOL advertisingTrackingEnabled;
+@property (strong, readonly) NSMutableArray<CTValidationResult*>* validationErrors;
 
-- (instancetype)initWithConfig:(CleverTapInstanceConfig *)config;
+- (instancetype)initWithConfig:(CleverTapInstanceConfig *)config andCleverTapID:(NSString *)cleverTapID;
 - (void)forceUpdateDeviceID:(NSString *)newDeviceID;
 - (void)forceNewDeviceID;
+- (void)forceUpdateCustomDeviceID:(NSString *)cleverTapID;
+- (BOOL)isErrorDeviceID;
 
 @end