+++ ExampleProject/platforms/ios/ExampleProject/Plugins/clevertap-cordova/CleverTapSDK.framework/PrivateHeaders/CTInAppDisplayViewController.h
@@ -1,10 +1,13 @@
 #import <UIKit/UIKit.h>
 #import "CTInAppNotification.h"
+#if !(TARGET_OS_TV)
+#import "CleverTapJSInterface.h"
+#endif
 
 @class CTInAppDisplayViewController;
 
 @protocol CTInAppNotificationDisplayDelegate <NSObject>
-- (void)handleNotificationCTA:(NSURL*)ctaURL forNotification:(CTInAppNotification*)notification fromViewController:(CTInAppDisplayViewController*)controller withExtras:(NSDictionary*)extras;
+- (void)handleNotificationCTA:(NSURL*)ctaURL buttonCustomExtras:(NSDictionary *)buttonCustomExtras forNotification:(CTInAppNotification*)notification fromViewController:(CTInAppDisplayViewController*)controller withExtras:(NSDictionary*)extras;
 - (void)notificationDidDismiss:(CTInAppNotification*)notification fromViewController:(CTInAppDisplayViewController*)controller;
 @optional
 - (void)notificationDidShow:(CTInAppNotification*)notification fromViewController:(CTInAppDisplayViewController*)controller;
@@ -17,8 +20,12 @@
 
 - (instancetype)init __unavailable;
 - (instancetype)initWithNotification:(CTInAppNotification*)notification;
+#if !(TARGET_OS_TV)
+- (instancetype)initWithNotification:(CTInAppNotification*)notification jsInterface:(CleverTapJSInterface *)jsInterface;
+#endif
 
 - (void)show:(BOOL)animated;
 - (void)hide:(BOOL)animated;
+- (BOOL)deviceOrientationIsLandscape;
 
 @end