+++ ExampleProject/platforms/ios/www/js/index.js
@@ -37,6 +37,7 @@ initialize: function() {
     // 'pause', 'resume', etc.
 onDeviceReady: function() {
     this.receivedEvent('deviceready');
+    CleverTap.setDebugLevel(3);
     CleverTap.registerPush();
     
     // Ionic example usage