+++ ExampleProject/plugins/clevertap-cordova/ExampleProject/platforms/ios/ExampleProject/Plugins/clevertap-cordova/CleverTapSDK.framework/PrivateHeaders/CTInboxSimpleMessageCell.h
@@ -1,0 +1,8 @@
+
+#import "CTInboxBaseMessageCell.h"
+
+@class FLAnimatedImageView;
+
+@interface CTInboxSimpleMessageCell : CTInboxBaseMessageCell
+
+@end