+++ ExampleProject/plugins/clevertap-cordova/src/android/CleverTapPlugin.java
@@ -45,10 +45,6 @@ import com.clevertap.android.sdk.EventDetail;
 import com.clevertap.android.sdk.UTMDetail;
 import com.clevertap.android.sdk.CTInboxListener;
 import com.clevertap.android.sdk.CTInboxStyleConfig;
-import com.clevertap.android.sdk.exceptions.CleverTapMetaDataNotFoundException;
-import com.clevertap.android.sdk.exceptions.CleverTapPermissionsNotSatisfied;
-import com.clevertap.android.sdk.exceptions.InvalidEventNameException;
-
 
 public class CleverTapPlugin extends CordovaPlugin implements SyncListener, InAppNotificationListener, CTInboxListener {
 
@@ -167,11 +163,36 @@ public class CleverTapPlugin extends CordovaPlugin implements SyncListener, InAp
             return true;
         }
 
-        // not required for Android here but handle as its in the JS interface
         else if (action.equals("setPushTokenAsString")) {
-            result = new PluginResult(PluginResult.Status.NO_RESULT);
-            result.setKeepCallback(true);
-            callbackContext.sendPluginResult(result);
+            final String token = args.getString(0);
+            cordova.getThreadPool().execute(new Runnable() {
+                public void run() {
+                    cleverTap.pushFcmRegistrationId(token,true);
+                    PluginResult _result = new PluginResult(PluginResult.Status.NO_RESULT);
+                    _result.setKeepCallback(true);
+                    callbackContext.sendPluginResult(_result);
+                }
+            });
+            return true;
+        }
+        
+        else if (action.equals("createNotification")) {
+           final String extras = args.getString(0);
+           JSONObject json = new JSONObject(extras);
+           Bundle bundle = new Bundle();
+           for (Iterator<String> entry = json.keys(); entry.hasNext(); ) {
+               String key = entry.next();
+               String str = json.optString(key);
+               bundle.putString(key,str);
+           }
+            cordova.getThreadPool().execute(new Runnable() {
+                public void run() {
+                    cleverTap.createNotification(cordova.getActivity().getApplicationContext(), bundle);
+                    PluginResult _result = new PluginResult(PluginResult.Status.NO_RESULT);
+                    _result.setKeepCallback(true);
+                    callbackContext.sendPluginResult(_result);
+                }
+            });
             return true;
         }
 
@@ -373,7 +394,7 @@ public class CleverTapPlugin extends CordovaPlugin implements SyncListener, InAp
             if (eventName != null) {
                 cordova.getThreadPool().execute(new Runnable() {
                     public void run() {
-                        cleverTap.event.push(eventName);
+                        cleverTap.pushEvent(eventName);
                         PluginResult _result = new PluginResult(PluginResult.Status.NO_RESULT);
                         _result.setKeepCallback(true);
                         callbackContext.sendPluginResult(_result);
@@ -420,7 +441,7 @@ public class CleverTapPlugin extends CordovaPlugin implements SyncListener, InAp
                 final HashMap<String, Object> props = _props;
                 cordova.getThreadPool().execute(new Runnable() {
                     public void run() {
-                        cleverTap.event.push(_eventName, props);
+                        cleverTap.pushEvent(_eventName, props);
                         PluginResult _result = new PluginResult(PluginResult.Status.NO_RESULT);
                         _result.setKeepCallback(true);
                         callbackContext.sendPluginResult(_result);
@@ -471,16 +492,10 @@ public class CleverTapPlugin extends CordovaPlugin implements SyncListener, InAp
                 final ArrayList<HashMap<String, Object>> items = _items;
                 cordova.getThreadPool().execute(new Runnable() {
                     public void run() {
-                        try {
-                            cleverTap.event.push(CleverTapAPI.CHARGED_EVENT, details, items);
-                            PluginResult _result = new PluginResult(PluginResult.Status.NO_RESULT);
-                            _result.setKeepCallback(true);
-                            callbackContext.sendPluginResult(_result);
-                        } catch (InvalidEventNameException e) {
-                            PluginResult _result = new PluginResult(PluginResult.Status.ERROR, e.getLocalizedMessage());
-                            _result.setKeepCallback(true);
-                            callbackContext.sendPluginResult(_result);
-                        }
+                        cleverTap.pushChargedEvent(details, items);
+                        PluginResult _result = new PluginResult(PluginResult.Status.NO_RESULT);
+                        _result.setKeepCallback(true);
+                        callbackContext.sendPluginResult(_result);
                     }
                 });
                 return true;
@@ -537,7 +552,7 @@ public class CleverTapPlugin extends CordovaPlugin implements SyncListener, InAp
             if (eventName != null) {
                 cordova.getThreadPool().execute(new Runnable() {
                     public void run() {
-                        double first = cleverTap.event.getFirstTime(eventName);
+                        double first = cleverTap.getFirstTime(eventName);
                         PluginResult _result = new PluginResult(PluginResult.Status.OK, (float) first);
                         _result.setKeepCallback(true);
                         callbackContext.sendPluginResult(_result);
@@ -555,7 +570,7 @@ public class CleverTapPlugin extends CordovaPlugin implements SyncListener, InAp
             if (eventName != null) {
                 cordova.getThreadPool().execute(new Runnable() {
                     public void run() {
-                        double lastTime = cleverTap.event.getLastTime(eventName);
+                        double lastTime = cleverTap.getLastTime(eventName);
                         PluginResult _result = new PluginResult(PluginResult.Status.OK, (float) lastTime);
                         _result.setKeepCallback(true);
                         callbackContext.sendPluginResult(_result);
@@ -573,7 +588,7 @@ public class CleverTapPlugin extends CordovaPlugin implements SyncListener, InAp
             if (eventName != null) {
                 cordova.getThreadPool().execute(new Runnable() {
                     public void run() {
-                        int num = cleverTap.event.getCount(eventName);
+                        int num = cleverTap.getCount(eventName);
                         PluginResult _result = new PluginResult(PluginResult.Status.OK, num);
                         _result.setKeepCallback(true);
                         callbackContext.sendPluginResult(_result);
@@ -591,7 +606,7 @@ public class CleverTapPlugin extends CordovaPlugin implements SyncListener, InAp
             if (eventName != null) {
                 cordova.getThreadPool().execute(new Runnable() {
                     public void run() {
-                        EventDetail details = cleverTap.event.getDetails(eventName);
+                        EventDetail details = cleverTap.getDetails(eventName);
                         try {
                             JSONObject jsonDetails = CleverTapPlugin.eventDetailsToJSON(details);
                             PluginResult _result = new PluginResult(PluginResult.Status.OK, jsonDetails);
@@ -614,7 +629,7 @@ public class CleverTapPlugin extends CordovaPlugin implements SyncListener, InAp
         else if (action.equals("getEventHistory")) {
             cordova.getThreadPool().execute(new Runnable() {
                 public void run() {
-                    Map<String, EventDetail> history = cleverTap.event.getHistory();
+                    Map<String, EventDetail> history = cleverTap.getHistory();
                     try {
                         JSONObject jsonDetails = CleverTapPlugin.eventHistoryToJSON(history);
                         PluginResult _result = new PluginResult(PluginResult.Status.OK, jsonDetails);
@@ -716,7 +731,7 @@ public class CleverTapPlugin extends CordovaPlugin implements SyncListener, InAp
                     public void run() {
                         try {
                             HashMap<String, Object> profile = formatProfile(_jsonProfile);
-                            cleverTap.profile.push(profile);
+                            cleverTap.pushProfile(profile);
                         } catch (Exception e) {
                             Log.d(LOG_TAG, "Error setting profile " + e.getLocalizedMessage());
                         }
@@ -786,7 +801,7 @@ public class CleverTapPlugin extends CordovaPlugin implements SyncListener, InAp
                 final JSONObject _jsonGraphUser = jsonGraphUser;
                 cordova.getThreadPool().execute(new Runnable() {
                     public void run() {
-                        cleverTap.profile.pushFacebookUser(_jsonGraphUser);
+                        cleverTap.pushFacebookUser(_jsonGraphUser);
                         PluginResult _result = new PluginResult(PluginResult.Status.NO_RESULT);
                         _result.setKeepCallback(true);
                         callbackContext.sendPluginResult(_result);
@@ -823,7 +838,7 @@ public class CleverTapPlugin extends CordovaPlugin implements SyncListener, InAp
                 final HashMap<String, Object> profile = _profile;
                 cordova.getThreadPool().execute(new Runnable() {
                     public void run() {
-                        cleverTap.profile.push(profile);
+                        cleverTap.pushProfile(profile);
                         PluginResult _result = new PluginResult(PluginResult.Status.NO_RESULT);
                         _result.setKeepCallback(true);
                         callbackContext.sendPluginResult(_result);
@@ -839,7 +854,7 @@ public class CleverTapPlugin extends CordovaPlugin implements SyncListener, InAp
                 cordova.getThreadPool().execute(new Runnable() {
                     public void run() {
                         PluginResult _result;
-                        Object prop = cleverTap.profile.getProperty(propertyName);
+                        Object prop = cleverTap.getProperty(propertyName);
 
                         if (prop instanceof JSONArray) {
                             JSONArray _prop = (JSONArray) prop;
@@ -895,7 +910,7 @@ public class CleverTapPlugin extends CordovaPlugin implements SyncListener, InAp
             if (key != null) {
                 cordova.getThreadPool().execute(new Runnable() {
                     public void run() {
-                        cleverTap.profile.removeValueForKey(key);
+                        cleverTap.removeValueForKey(key);
                         PluginResult _result = new PluginResult(PluginResult.Status.NO_RESULT);
                         _result.setKeepCallback(true);
                         callbackContext.sendPluginResult(_result);
@@ -944,7 +959,7 @@ public class CleverTapPlugin extends CordovaPlugin implements SyncListener, InAp
 
                     cordova.getThreadPool().execute(new Runnable() {
                         public void run() {
-                            cleverTap.profile.setMultiValuesForKey(_key, _values);
+                            cleverTap.setMultiValuesForKey(_key, _values);
                             PluginResult _result = new PluginResult(PluginResult.Status.NO_RESULT);
                             _result.setKeepCallback(true);
                             callbackContext.sendPluginResult(_result);
@@ -995,7 +1010,7 @@ public class CleverTapPlugin extends CordovaPlugin implements SyncListener, InAp
 
                     cordova.getThreadPool().execute(new Runnable() {
                         public void run() {
-                            cleverTap.profile.addMultiValuesForKey(_key, _values);
+                            cleverTap.addMultiValuesForKey(_key, _values);
                             PluginResult _result = new PluginResult(PluginResult.Status.NO_RESULT);
                             _result.setKeepCallback(true);
                             callbackContext.sendPluginResult(_result);
@@ -1045,7 +1060,7 @@ public class CleverTapPlugin extends CordovaPlugin implements SyncListener, InAp
 
                     cordova.getThreadPool().execute(new Runnable() {
                         public void run() {
-                            cleverTap.profile.removeMultiValuesForKey(_key, _values);
+                            cleverTap.removeMultiValuesForKey(_key, _values);
                             PluginResult _result = new PluginResult(PluginResult.Status.NO_RESULT);
                             _result.setKeepCallback(true);
                             callbackContext.sendPluginResult(_result);
@@ -1088,7 +1103,7 @@ public class CleverTapPlugin extends CordovaPlugin implements SyncListener, InAp
 
                 cordova.getThreadPool().execute(new Runnable() {
                     public void run() {
-                        cleverTap.profile.addMultiValueForKey(_key, _value);
+                        cleverTap.addMultiValueForKey(_key, _value);
                         PluginResult _result = new PluginResult(PluginResult.Status.NO_RESULT);
                         _result.setKeepCallback(true);
                         callbackContext.sendPluginResult(_result);
@@ -1126,7 +1141,7 @@ public class CleverTapPlugin extends CordovaPlugin implements SyncListener, InAp
 
                 cordova.getThreadPool().execute(new Runnable() {
                     public void run() {
-                        cleverTap.profile.removeMultiValueForKey(_key, _value);
+                        cleverTap.removeMultiValueForKey(_key, _value);
                         PluginResult _result = new PluginResult(PluginResult.Status.NO_RESULT);
                         _result.setKeepCallback(true);
                         callbackContext.sendPluginResult(_result);
@@ -1139,7 +1154,7 @@ public class CleverTapPlugin extends CordovaPlugin implements SyncListener, InAp
         else if (action.equals("sessionGetTimeElapsed")) {
             cordova.getThreadPool().execute(new Runnable() {
                 public void run() {
-                    int time = cleverTap.session.getTimeElapsed();
+                    int time = cleverTap.getTimeElapsed();
                     PluginResult _result = new PluginResult(PluginResult.Status.OK, time);
                     _result.setKeepCallback(true);
                     callbackContext.sendPluginResult(_result);
@@ -1151,7 +1166,7 @@ public class CleverTapPlugin extends CordovaPlugin implements SyncListener, InAp
         else if (action.equals("sessionGetTotalVisits")) {
             cordova.getThreadPool().execute(new Runnable() {
                 public void run() {
-                    int count = cleverTap.session.getTotalVisits();
+                    int count = cleverTap.getTotalVisits();
                     PluginResult _result = new PluginResult(PluginResult.Status.OK, count);
                     _result.setKeepCallback(true);
                     callbackContext.sendPluginResult(_result);
@@ -1163,7 +1178,7 @@ public class CleverTapPlugin extends CordovaPlugin implements SyncListener, InAp
         else if (action.equals("sessionGetScreenCount")) {
             cordova.getThreadPool().execute(new Runnable() {
                 public void run() {
-                    int count = cleverTap.session.getScreenCount();
+                    int count = cleverTap.getScreenCount();
                     PluginResult _result = new PluginResult(PluginResult.Status.OK, count);
                     _result.setKeepCallback(true);
                     callbackContext.sendPluginResult(_result);
@@ -1175,7 +1190,7 @@ public class CleverTapPlugin extends CordovaPlugin implements SyncListener, InAp
         else if (action.equals("sessionGetPreviousVisitTime")) {
             cordova.getThreadPool().execute(new Runnable() {
                 public void run() {
-                    int time = cleverTap.session.getPreviousVisitTime();
+                    int time = cleverTap.getPreviousVisitTime();
                     PluginResult _result = new PluginResult(PluginResult.Status.OK, time);
                     _result.setKeepCallback(true);
                     callbackContext.sendPluginResult(_result);
@@ -1187,7 +1202,7 @@ public class CleverTapPlugin extends CordovaPlugin implements SyncListener, InAp
         else if (action.equals("sessionGetUTMDetails")) {
             cordova.getThreadPool().execute(new Runnable() {
                 public void run() {
-                    UTMDetail details = cleverTap.session.getUTMDetails();
+                    UTMDetail details = cleverTap.getUTMDetails();
                     try {
                         JSONObject jsonDetails = CleverTapPlugin.utmDetailsToJSON(details);
                         PluginResult _result = new PluginResult(PluginResult.Status.OK, jsonDetails);